quell 10.5 errors, CG interfaces taking GLint pointers
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/v2_0branch/bzflag@16144 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/platform/MacOSX/MacWindow.cxx
@@ -103,7 +103,7 @@ class GLContext
 
     bool Init(u_int32_t display_id, const CGRect& display_rect) {
       CGLPixelFormatObj pixel_format;
-      long num_pixel_formats;
+      GLint num_pixel_formats;
 
       CGLError err = CGLChoosePixelFormat(GetPixelFormat(display_id, 24),
 	  &pixel_format, &num_pixel_formats);
@@ -185,7 +185,7 @@ class GLContext
     }
 
     void SetVBLSynch(bool synch) {
-      long params[] = { synch ? 1 : 0 };
+      GLint params[] = { synch ? 1 : 0 };
       CGLSetParameter(cgl_context, kCGLCPSwapInterval, params);
 
       settings.VBL_Synch = synch;
